refactor(util, movement): MoveStrategy 인터페이스 도입 및 service 구조를 util로 개선
- CarMoveManager를 MoveStrategy 인터페이스 기반으로 리팩토링하고 CarMoveStrategy로 구현
- CarController와 Application에서 MoveStrategy 인터페이스를 사용해 의존성 주입
- 기존 service 패키지의 클래스를 util 패키지 내로 재구성
- 이동 전략 변경 시 유연성과 확장성을 확보
