+++ src/main/java/racingcar/Application.java
@@ -1,18 +1,19 @@
 package racingcar;
 
 import racingcar.controller.CarController;
-import racingcar.service.CarMoveManager;
-import racingcar.service.InputSplitter;
-import racingcar.service.RandomNumberGenerator;
+import racingcar.util.generator.RandomNumberGenerator;
+import racingcar.util.movement.CarMoveStrategy;
+import racingcar.util.movement.MoveStrategy;
+import racingcar.util.parser.InputSplitter;
 
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        CarMoveManager carMoveManager = new CarMoveManager();
+        MoveStrategy moveStrategy = new CarMoveStrategy();
         InputSplitter inputSplitter = new InputSplitter();
         RandomNumberGenerator randomNumberGenerator = new RandomNumberGenerator();
 
-        CarController carController = new CarController(carMoveManager, inputSplitter, randomNumberGenerator);
+        CarController carController = new CarController(moveStrategy, inputSplitter, randomNumberGenerator);
 
         carController.play();
     }
