+++ src/main/java/racingcar/controller/CarController.java
@@ -3,22 +3,22 @@ package racingcar.controller;
 import java.util.List;
 import racingcar.model.Car;
 import racingcar.model.Cars;
-import racingcar.service.CarMoveManager;
-import racingcar.service.InputSplitter;
-import racingcar.service.RandomNumberGenerator;
+import racingcar.util.generator.RandomNumberGenerator;
+import racingcar.util.movement.MoveStrategy;
+import racingcar.util.parser.InputSplitter;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
 public class CarController {
-    private final CarMoveManager carMoveManager;
+    private final MoveStrategy moveStrategy;
     private final InputSplitter inputSplitter;
     private final RandomNumberGenerator randomNumberGenerator;
     private final InputView inputView;
     private final OutputView outputView;
 
-    public CarController(CarMoveManager carMoveManager, InputSplitter inputSplitter,
+    public CarController(MoveStrategy moveStrategy, InputSplitter inputSplitter,
                          RandomNumberGenerator randomNumberGenerator) {
-        this.carMoveManager = carMoveManager;
+        this.moveStrategy = moveStrategy;
         this.inputSplitter = inputSplitter;
         this.randomNumberGenerator = randomNumberGenerator;
         this.inputView = InputView.getInstance();
@@ -40,7 +40,7 @@ public class CarController {
         for (int i = 0; i < count; i++) {
             for (Car car : cars.getCars()) {
                 int number = randomNumberGenerator.gernerateRandomNumber();
-                car.move(carMoveManager.canMove(number));
+                car.move(moveStrategy.canMove(number));
             }
             outputView.showRoundResult(cars);
         }
