+++ src/main/java/racingcar/util/movement/CarMoveStrategy.java
@@ -1,8 +1,9 @@
-package racingcar.service;
+package racingcar.util.movement;
 
-public class CarMoveManager {
+public class CarMoveStrategy implements MoveStrategy {
     private static final int MOVE_THRESHOLD = 4;
 
+    @Override
     public boolean canMove(int number) {
         return number >= MOVE_THRESHOLD;
     }
