+++ src/main/java/racingcar/util/movement/MoveStrategy.java
@@ -1,0 +1,5 @@
+package racingcar.util.movement;
+
+public interface MoveStrategy {
+    boolean canMove(int number);
+}
